fix leftovers
