+++ apps/segment/src/modules/db/segment-apl-repository-factory.ts
@@ -1,16 +1,16 @@
 import { BaseError } from "@/errors";
 
-import { SegmentConfigRepository } from "./segment-apl-repository";
+import { SegmentAPLRepository } from "./segment-apl-repository";
 import { segmentMainTable, SegmentMainTableEntityFactory } from "./segment-main-table";
 
 export class SegmentAPLRepositoryFactory {
   static RepositoryCreationError = BaseError.subclass("RepositoryCreationError");
 
-  static create(): SegmentConfigRepository {
+  static create(): SegmentAPLRepository {
     try {
       const segmentAPLEntity = SegmentMainTableEntityFactory.createAPLEntity(segmentMainTable);
 
-      return new SegmentConfigRepository({ segmentAPLEntity });
+      return new SegmentAPLRepository({ segmentAPLEntity });
     } catch (error) {
       throw new SegmentAPLRepositoryFactory.RepositoryCreationError(
         "Failed to create DynamoDB APL repository",
