+++ apps/segment/src/modules/db/segment-apl-repository.ts
@@ -9,7 +9,7 @@ import { SegmentAPLEntityType, SegmentMainTable } from "@/modules/db/segment-mai
 import { SegmentAPLMapper } from "./segment-apl-mapper";
 import { APLRepository } from "./types";
 
-export class SegmentConfigRepository implements APLRepository {
+export class SegmentAPLRepository implements APLRepository {
   private logger = createLogger("SegmentAPLRepository");
 
   private segmentAPLMapper = new SegmentAPLMapper();
@@ -37,7 +37,7 @@ export class SegmentConfigRepository implements APLRepository {
         })
         .send(),
       (error) =>
-        new SegmentConfigRepository.ReadEntityError("Failed to read APL entity", { cause: error }),
+        new SegmentAPLRepository.ReadEntityError("Failed to read APL entity", { cause: error }),
     );
 
     if (getEntryResult.isErr()) {
@@ -64,7 +64,7 @@ export class SegmentConfigRepository implements APLRepository {
         .item(this.segmentAPLMapper.authDataToDynamoPutEntity(args.authData))
         .send(),
       (error) =>
-        new SegmentConfigRepository.WriteEntityError("Failed to write APL entity", {
+        new SegmentAPLRepository.WriteEntityError("Failed to write APL entity", {
           cause: error,
         }),
     );
@@ -92,7 +92,7 @@ export class SegmentConfigRepository implements APLRepository {
         })
         .send(),
       (error) =>
-        new SegmentConfigRepository.DeleteEntityError("Failed to delete APL entity", {
+        new SegmentAPLRepository.DeleteEntityError("Failed to delete APL entity", {
           cause: error,
         }),
     );
@@ -119,7 +119,7 @@ export class SegmentConfigRepository implements APLRepository {
         })
         .send(),
       (error) =>
-        new SegmentConfigRepository.ScanEntityError("Failed to scan APL entities", {
+        new SegmentAPLRepository.ScanEntityError("Failed to scan APL entities", {
           cause: error,
         }),
     );
